% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc_core}
\alias{mcmc_core}
\title{Perform Markov Chain Monte Carlo (MCMC) sampling using
Metropolis-Hasting algorithm.}
\usage{
mcmc_core(mut_data, mut_rate, num_iters = 1000, burning = 0,
  mcmc_tuning, tg_lambda, mcmc_output, geneset_size, seed = -1,
  previous_stats = NULL, debug = -1)
}
\arguments{
\item{mut_data}{Mutation table created using MADGIC algorithm.}

\item{mut_rate}{Mutation rate table create uisng MADGIC algorithm.}

\item{num_iters}{Number of MCMC samples.}

\item{burning}{Number of Burn-in samples.}

\item{mcmc_tuning}{MCMC tuning parameter.}

\item{tg_lambda}{lambda.}

\item{geneset_size}{Size of the gene sets to sample.}

\item{seed}{Seed for RGN.}

\item{previous_stats}{Recovery from the previous MCMC chain file.}
}
\value{
Tab-delimited file contain gene sets and TG scores
}
\description{
Perform Markov Chain Monte Carlo (MCMC) sampling using
Metropolis-Hasting algorithm.
}
