% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc_tuning}
\alias{mcmc_tuning}
\title{Select tuning parameters for Markov Chain Monte Carlo (MCMC) sampling
based on the size of the gene sets}
\usage{
mcmc_tuning(mut_data, mut_rate, num_iters = 1000, burning = 0,
  mcmc_tuning, tg_lambda, seed = -1, range = c(2:3))
}
\arguments{
\item{mut_data}{Mutation table created using MADGIC algorithm.}

\item{mut_rate}{Mutation rate table create uisng MADGIC algorithm.}

\item{num_iters}{Number of MCMC samples.}

\item{burning}{Number of Burn-in samples.}

\item{mcmc_tuning}{MCMC initial tuning parameter, .}

\item{tg_lambda}{lambda}

\item{seed}{Seed for RGN.}
}
\value{
tuning parameter after the selection
}
\description{
Select tuning parameters for Markov Chain Monte Carlo (MCMC) sampling
based on the size of the gene sets
}
