% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mescan_C.R
\name{scan_genesets}
\alias{scan_genesets}
\title{Return TG scores for the gene sets combinations from \code{gs_idx},
with size of \code{n}.}
\usage{
scan_genesets(gs_idx, n = 2, mut_data, rate_data, lambda)
}
\arguments{
\item{gs_idx}{A vector of genes index}

\item{n}{The size of the gene sets}

\item{mut_data}{Mutation table created using MADGIC algorithm.}

\item{lambda}{lambda.}

\item{mut_rate}{Mutation rate table create uisng MADGIC algorithm.}
}
\value{
A list contains the gene set with maximun TG score, the maximum TG score and all the scanned gene sets
}
\description{
Return TG scores for the gene sets combinations from \code{gs_idx},
with size of \code{n}.
}
